переделал коментарии и переместил их в другое место
