+++ .idea/misc.xml
@@ -14,7 +14,7 @@
         <entry key="..\:/Programming/SalesApp/app/src/main/res/drawable/ic_prost.xml" value="0.2703125" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/drawable/ic_site_b.xml" value="0.2703125" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout-v26/activity_main.xml" value="0.25052083333333336" />
-        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_about_us.xml" value="0.536" />
+        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_about_us.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_cart_shop_page.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_error.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_intro.xml" value="0.25416666666666665" />
@@ -22,17 +22,19 @@
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_main.xml" value="0.264" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_main_comm.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_matching.xml" value="0.2542991141219385" />
-        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_shop_page.xml" value="0.4" />
+        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_shop_page.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_sign_in.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/activity_splash.xml" value="0.25052083333333336" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/cart_item.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/category_item.xml" value="0.25416666666666665" />
+        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/custom_dialog.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/intro_four.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/intro_one.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/intro_three.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/intro_two.xml" value="0.264" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/item.xml" value="0.1" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/shop_item.xml" value="0.25025641025641027" />
+        <entry key="..\:/Programming/SalesApp/app/src/main/res/layout/user_message.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/menu/main_menu.xml" value="0.25416666666666665" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/mipmap-anydpi-v26/ic_launcher.xml" value="0.2703125" />
         <entry key="..\:/Programming/SalesApp/app/src/main/res/mipmap-anydpi-v26/ic_launcher_round.xml" value="0.2703125" />
