+++ app/src/main/java/com/example/salesapp/activities/AboutUsActivity.java
@@ -17,18 +17,20 @@ import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.example.salesapp.Matching;
 import com.example.salesapp.R;
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class AboutUsActivity extends AppCompatActivity {
 
     TextView main_scene, coming_soon_scene;
     ImageView ico;
+    FloatingActionButton commentsLink;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_about_us);
+        getSupportActionBar().hide();
 
         main_scene = findViewById(R.id.main_scene);
         main_scene.setOnClickListener(new View.OnClickListener() {
@@ -57,9 +59,20 @@ public class AboutUsActivity extends AppCompatActivity {
 
             }
         });
+        commentsLink =findViewById(R.id.commentsFloatButton);
+
+        commentsLink.setOnClickListener(v -> {
+            //v.startAnimation(AnimationUtils.loadAnimation(ShopPage.this, R.anim.anim_item));
+            openCommentsActivity();
+
+        });
 
     }
 
+    private void openCommentsActivity() {
+        startActivity(new Intent(this, SignInAct.class));
+
+    }
     public void openGit(View v){
         Vibrator v1 = (Vibrator) AboutUsActivity.this.getSystemService(Context.VIBRATOR_SERVICE);
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
