+++ app/src/main/java/com/example/salesapp/activities/ErrorPage.java
@@ -1,4 +1,4 @@
-package com.example.salesapp;
+package com.example.salesapp.activities;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -11,10 +11,14 @@ import android.widget.VideoView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.salesapp.R;
+
 public class ErrorPage extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_error);
+        getSupportActionBar().hide();
+
         ImageView link_img = findViewById(R.id.imageView6);
         link_img.setOnClickListener(v -> {
             startActivity(new Intent(Settings.ACTION_WIFI_SETTINGS));
