+++ app/src/main/java/com/example/salesapp/activities/IntroActivity.java
@@ -31,6 +31,7 @@ public class IntroActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_intro);
+        getSupportActionBar().hide();
 
         introPref = new IntroPref(this);
         if (!introPref.isFirstTimeLaunch()) {
