+++ app/src/main/java/com/example/salesapp/activities/Matching.java
@@ -1,4 +1,4 @@
-package com.example.salesapp;
+package com.example.salesapp.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -16,8 +16,7 @@ import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 import android.widget.TextView;
 
-import com.example.salesapp.activities.AboutUsActivity;
-import com.example.salesapp.activities.MainActivity;
+import com.example.salesapp.R;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 public class Matching extends AppCompatActivity {
@@ -29,6 +28,7 @@ public class Matching extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_matching);
+        getSupportActionBar().hide();
 
         main_scene = findViewById(R.id.main_scene);
