+++ app/src/main/res/layout/activity_main_comm.xml
@@ -41,16 +41,14 @@
         android:layout_height="52dp"
         android:layout_marginEnd="16dp"
         android:layout_marginBottom="52dp"
-        android:backgroundTint="@color/item_bg"
 
         android:clickable="true"
-        android:tint="#9C27B0"
+        app:backgroundTint="@color/cat"
         app:fabSize="auto"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toEndOf="@+id/edMessage"
-        app:rippleColor="#2196F3"
         app:srcCompat="@drawable/ic_baseline_wrap_text_24"
         tools:ignore="SpeakableTextPresentCheck" />
