+++ app/src/main/res/layout/activity_shop_page.xml
@@ -160,17 +160,6 @@
         app:layout_constraintVertical_bias="0.0"
         tools:listitem="@layout/item" />
 
-    <com.google.android.material.floatingactionbutton.FloatingActionButton
-        android:id="@+id/commentsFloatButton"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginEnd="16dp"
-        android:layout_marginBottom="16dp"
-        android:clickable="true"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:srcCompat="@drawable/ic_baseline_comment_24"
-        tools:ignore="SpeakableTextPresentCheck" />
 
 
     <!--    <ImageView-->
